+++ spec/lib/votd/base_spec.rb
@@ -35,26 +35,39 @@ describe "Votd::Base" do
 
   describe ".to_html" do
     it "returns a HTML version" do
-      votd.to_html.should == File.read(fixture("base.html"))
+      votd.to_html.should == read_fixture("base/base.html")
     end
   end
 
   describe ".custom_html" do
-    it "overrides the default .to_html formatting" do
-      votd.custom_html do |votd|
+    it "returns custom HTML and overrides the default .to_html formatting" do
+      html_from_block = votd.custom_html do |votd|
         "<p>#{votd.reference}|#{votd.text}|#{votd.version}</p>"
       end
-      votd.to_html.should == "<p>John 3:16|For God so loved the world, that he gave his only begotten Son, that whosoever believeth in him should not perish, but have everlasting life.|KJV</p>"
+      desired_output = read_fixture("base/base_custom.html")
+      html_from_block.should == desired_output
+      votd.to_html.should    == desired_output
     end
   end
 
   describe ".to_text" do
-    it "returns a Text-formatted version" do
-      votd.to_text.should == File.read(fixture("base.txt"))
+    it "returns a text-formatted version" do
+      votd.to_text.should == read_fixture("base/base.txt")
     end
 
     it "is aliased to .to_s" do
-      votd.to_s.should == File.read(fixture("base.txt"))
+      votd.to_s.should == read_fixture("base/base.txt")
+    end
+  end
+
+  describe ".custom_text" do
+    it "overrides the default.to_text formatting" do
+      text_from_block = votd.custom_text do |votd|
+        "#{votd.reference}|#{votd.text}|#{votd.version}"
+      end
+      desired_output = read_fixture("base/base_custom.txt")
+      text_from_block.should == desired_output
+      votd.to_text.should    == desired_output
     end
   end
 end