+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 * CURRENT ITERATION (2.1.0.pre)
-  * Add `custom_html` method that takes a block of custom formatted
+  * [Add] `custom_html` method that takes a block of custom formatted
     HTML to override the `to_html` method.
+  * [Add] `to_text` method that returns a text-formatted version of the VotD.
+    Aliased to `to_s`.
+  * [Add] `custom_text` method that takes a block of custom formatted text
+    to override the `to_text` method.
 * March 30, 2010 2.0.0 release
   * New `copyright` attribute containing any copyright information provided
     by the VotD service. This may will cause any existing use of